Remove node-rpc, move to control rpc
