+++ dashboard/src/front/components/fluence-nodes-list/index.tsx
