--- effects/node-rpc/src/main/scala/fluence/effects/node/rpc/RpcError.scala
@@ -1,5 +1,0 @@
-package fluence.effects.node.rpc
-
-import fluence.effects.EffectError
-
-sealed trait RpcError extends EffectError
